refactor: move postsGrid folder
