+++ src/entities/post/ui/postsGrid/PostsGrid.module.scss
