+++ src/features/post/ui/postsGridWithInfiniteScroll/PostsGridWithInfinteScroll.tsx
@@ -1,7 +1,8 @@
 'use client'
+
 import { useAuth } from '@/features/auth/hooks/useAuth'
 import { useGetAllPostsInfiniteQuery } from '@/features/post/api/postsApi'
-import { PostsGrid } from '@/features/post/ui/createPost/postsGrid/PostsGrid'
+import { PostsGrid } from '@/entities/post/ui/postsGrid/PostsGrid'
 import { Loader } from '@/shared/ui/loader/Loader'
 import { skipToken } from '@reduxjs/toolkit/query'
 import { useCallback, useEffect, useRef } from 'react'
